+++ src/models/userModel.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const { DataTypes } = require('sequelize');
+const { sequelize } = require('../db/db');
+
+const User = sequelize.define('User', {
+  name: {
+    type: DataTypes.STRING,
+    allowNull: false,
+  },
+}, {
+  tableName: 'users',
+  updatedAt: false,
+  createdAt: false,
+});
+
+module.exports = { User };
