+++ src/services/users.service.js
@@ -1,18 +1,6 @@
 'use strict';
 
-const { DataTypes } = require('sequelize');
-const { sequelize } = require('../db/db');
-
-const User = sequelize.define('User', {
-  name: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-}, {
-  tableName: 'users',
-  updatedAt: false,
-  createdAt: false,
-});
+const { User } = require('../models/userModel');
 
 const getAll = () => {
   return User.findAll();
@@ -46,5 +34,4 @@ module.exports = {
   create,
   update,
   remove,
-  User,
 };
